# event_go/internal/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Soulmarker Event API",
    "version": "1.0.0",
    "description": "Versioned API contract for authentication, organization authorization, organizers, events, registrations, admissions, checkins, discussions and tickets. Legacy /api routes remain available for one compatibility cycle."
  },
  "servers": [{ "url": "/api/v1" }],
...
  "paths": {
...
    "/me/registrations": {
      "get": {
        "operationId": "listMyRegistrations",
        "tags": ["Me"],
        "summary": "List registrations owned by the current user",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/events/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/EventID" }],
      "get": {
        "operationId": "getEvent",
        "tags": ["Events"],
        "summary": "Get an event",
        "responses": { "200": { "$ref": "#/components/responses/OK" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "put": {
        "operationId": "updateEvent",
        "tags": ["Events"],
        "summary": "Update an event",
        "security": [{ "AdminToken": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/UpdateEventRequest" },
        "responses": { "200": { "$ref": "#/components/responses/OK" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "delete": {
        "operationId": "deleteEvent",
        "tags": ["Events"],
        "summary": "Delete an event and dependent records",
        "security": [{ "AdminToken": [] }],
        "responses": { "200": { "$ref": "#/components/responses/OK" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "AdminToken": { "type": "apiKey", "in": "header", "name": "X-Admin-Token", "description": "Platform administrator credential; it does not grant an organization membership or tenant capability." }
    },
    "parameters": {
      "Page": { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
      "PageSize": { "name": "page_size", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } },
...
      "EventID": { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
...
    },
    "requestBodies": {
...
      "UpdateEventRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateEventRequest" } } } },
...
    },
    "responses": {
      "OK": { "description": "Successful response", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/APIResponse" } } } },
...
      "BadRequest": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Unauthorized": { "description": "Authentication failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
      "NotFound": { "description": "Resource not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
    },
    "schemas": {
      "APIResponse": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": { "type": "integer" },
          "error_code": { "type": "string", "enum": ["ADMIN_AUTH_INVALID", "ADMISSION_ALREADY_CHECKED_IN", "ADMISSION_NOT_FOUND", "ADMISSION_REVOKED", "API_ROUTE_NOT_FOUND", "CANCELLATION_DEADLINE_EXCEEDED", "CONTENT_ALREADY_REMOVED", "CONTENT_ALREADY_VISIBLE", "CONTENT_REPORT_ALREADY_RESOLVED", "CONTENT_REPORT_NOT_ALLOWED", "CONTENT_REPORT_NOT_FOUND", "EVENT_CAPACITY_FULL", "EVENT_HAS_ADMISSIONS", "EVENT_NOT_FOUND", "EVENT_NOT_PUBLISHED", "INTERNAL_ERROR", "INVALID_CREDENTIALS", "INVALID_JSON", "METHOD_NOT_ALLOWED", "NOTIFICATION_NOT_FOUND", "ORGANIZATION_ACCESS_DENIED", "ORGANIZATION_INVITATION_EXISTS", "ORGANIZATION_INVITATION_INVALID", "ORGANIZATION_MEMBER_CHANGE_DENIED", "ORGANIZATION_MEMBER_EXISTS", "ORGANIZATION_MEMBER_NOT_FOUND", "ORGANIZATION_OWNER_TRANSFER_DENIED", "ORGANIZATION_SLUG_IN_USE", "ORGANIZER_NOT_FOUND", "PARTICIPATION_REQUIRED", "PASSWORD_RESET_TOKEN_INVALID", "POST_NOT_FOUND", "RECOVERY_EMAIL_ALREADY_BOUND", "RECOVERY_EMAIL_IN_USE", "RECOVERY_EMAIL_RATE_LIMITED", "RECOVERY_EMAIL_TOKEN_INVALID", "REGISTRATION_DUPLICATE", "REGISTRATION_NOT_FOUND", "REPLY_NOT_FOUND", "REQUEST_TOO_LARGE", "TICKET_NOT_FOUND", "TICKET_SOLD_OUT", "USER_ALREADY_EXISTS", "USER_AUTH_REQUIRED", "USER_TOKEN_INVALID", "VALIDATION_ERROR"] },
          "message": { "type": "string" },
          "data": {},
          "total": { "type": "integer", "minimum": 0 },
          "page": { "type": "integer", "minimum": 1 },
          "page_size": { "type": "integer", "minimum": 1, "maximum": 100 }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "error_code", "message"],
        "properties": {
          "code": { "type": "integer", "minimum": 400, "maximum": 599 },
          "error_code": { "type": "string", "enum": ["ADMIN_AUTH_INVALID", "ADMISSION_ALREADY_CHECKED_IN", "ADMISSION_NOT_FOUND", "ADMISSION_REVOKED", "API_ROUTE_NOT_FOUND", "CANCELLATION_DEADLINE_EXCEEDED", "CONTENT_ALREADY_REMOVED", "CONTENT_ALREADY_VISIBLE", "CONTENT_REPORT_ALREADY_RESOLVED", "CONTENT_REPORT_NOT_ALLOWED", "CONTENT_REPORT_NOT_FOUND", "EVENT_CAPACITY_FULL", "EVENT_HAS_ADMISSIONS", "EVENT_NOT_FOUND", "EVENT_NOT_PUBLISHED", "INTERNAL_ERROR", "INVALID_CREDENTIALS", "INVALID_JSON", "METHOD_NOT_ALLOWED", "NOTIFICATION_NOT_FOUND", "ORGANIZATION_ACCESS_DENIED", "ORGANIZATION_INVITATION_EXISTS", "ORGANIZATION_INVITATION_INVALID", "ORGANIZATION_MEMBER_CHANGE_DENIED", "ORGANIZATION_MEMBER_EXISTS", "ORGANIZATION_MEMBER_NOT_FOUND", "ORGANIZATION_OWNER_TRANSFER_DENIED", "ORGANIZATION_SLUG_IN_USE", "ORGANIZER_NOT_FOUND", "PARTICIPATION_REQUIRED", "PASSWORD_RESET_TOKEN_INVALID", "POST_NOT_FOUND", "RECOVERY_EMAIL_ALREADY_BOUND", "RECOVERY_EMAIL_IN_USE", "RECOVERY_EMAIL_RATE_LIMITED", "RECOVERY_EMAIL_TOKEN_INVALID", "REGISTRATION_DUPLICATE", "REGISTRATION_NOT_FOUND", "REPLY_NOT_FOUND", "REQUEST_TOO_LARGE", "TICKET_NOT_FOUND", "TICKET_SOLD_OUT", "USER_ALREADY_EXISTS", "USER_AUTH_REQUIRED", "USER_TOKEN_INVALID", "VALIDATION_ERROR"] },
          "message": { "type": "string" }
        }
      },
...
      "UpdateEventRequest": { "type": "object", "additionalProperties": false, "properties": { "organizer_id": { "type": "integer", "format": "int64", "minimum": 1 }, "title": { "type": "string" }, "description": { "type": "string" }, "cover_url": { "type": "string", "format": "uri-reference" }, "event_time": { "type": "string", "format": "date-time" }, "location": { "type": "string" }, "capacity": { "type": "integer", "minimum": 1 }, "price": { "type": "number", "minimum": 0 }, "status": { "type": "string", "enum": ["draft", "published", "cancelled", "ended"] } } },
...
    }
  }
}